backoffice/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.10.0
requests==2.31.0
gunicorn==21.2.0
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.5.1
